Preserve most of the original outline
When `outline` is set to `none` or `0` it's impossible to revert back to the UA default. By just setting `outline-width` to `0` I can at least set it back to `3px` which gives the correct default border in Chrome. I believe Firefox uses a 2px border so it might be a little chunky but better than having to totally recreate the outline styles for different platforms.  Thanks to @atotic for the tip!